* Merge SVN revision r1240 from branch devel into stable.
git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/stable@1241 688f6386-cf32-0410-9ead-a596f0484e40

+++ bin/mindump-repaint-labels.sh
@@ -56,31 +56,32 @@ then
     exit 1
 fi
 
-# Create list with all labels (reverse sorted)
-$NM -r -n -C $LINUXSYS | uniq > $NMFILE
+# Check installed bc
+if ! which bc >/dev/null
+then
+    echo "bc: Not installed or not found." >&2
+    exit 1
+fi
+
+# Create list with all labels (reverse sorted), remove ".text" and ".bss"
+$NM -r -n -t d -C $LINUXSYS | grep -v -e ' \.' | uniq > $NMFILE
 if [ ! -s $NMFILE ]
 then
     rm $NMFILE
     exit 1
 fi
 
-# Convert address into decimal
-( echo "ibase=16"; cut -d ' ' -f 1 < $NMFILE | tr [[:lower:]] [[:upper:]] ) | bc > $NMFILE.numbers
-cut -d ' ' -f 3 < $NMFILE > $NMFILE.names
-paste $NMFILE.numbers $NMFILE.names > $NMFILE.dec
-rm $NMFILE $NMFILE.numbers $NMFILE.names
-
 # Search start address and label name for given offset
 search_label()
 {
-    while read address name
+    while read address type name
     do
 	if [ $address -le $1 ]
 	then
 	    echo "$address $name"
 	    return
 	fi
-    done < $NMFILE.dec
+    done < $NMFILE
 }
 
 # Serach and replace all "linux+0x..." with label names and offsets
@@ -88,7 +89,7 @@ start=false
 sed -e 's/\r//' |\
 while read s1 s2 s3 s4 s5 name
 do
-    if [ "$s1" = "STACK_TEXT:" ]
+    if [ "$s1" = "STACK_TEXT:" -o "$s1" = "ChildEBP" ]
     then
 	start=true
     elif [ -z "$s1" ]
@@ -96,7 +97,7 @@ do
 	start=false
     elif $start
     then
-	offset=`echo "$name" | sed -n -r -e 's/^linux\+0x(.+)$/\1/p' | tr [[:lower:]] [[:upper:]]`
+	offset=`echo "$name" | sed -n -e 's/^linux+0x\(.*\)$/\1/p' | tr [[:lower:]] [[:upper:]]`
 	if [ -n "$offset" ]
 	then
 	    # Add offset 0x10000 and convert from hex into decimal
@@ -113,4 +114,4 @@ do
 done
 
 # Remove temp file
-rm $NMFILE.dec
+rm $NMFILE
